+++ packages/fx-core/src/common/m365/launchHelper.ts
@@ -11,6 +11,7 @@ import { serviceEndpoint, serviceScope } from "./serviceConstant";
 import { assembleError } from "../../error/common";
 import { HubTypes } from "../../question/other";
 import { ErrorContextMW } from "../../core/globalVars";
+import { hooks } from "@feathersjs/hooks";
 
 export class LaunchHelper {
   private readonly m365TokenProvider: M365TokenProvider;
@@ -20,7 +21,7 @@ export class LaunchHelper {
     this.m365TokenProvider = m365TokenProvider;
     this.logger = logger;
   }
-  @ErrorContextMW({ component: "LaunchHelper" })
+  @hooks([ErrorContextMW({ component: "LaunchHelper" })])
   public async getLaunchUrl(
     hub: HubTypes,
     teamsAppId: string,
