+++ packages/fx-core/src/component/coordinator/index.ts
@@ -1,29 +1,28 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
+import { hooks } from "@feathersjs/hooks/lib";
+import {
+  Colors,
+  Context,
+  CreateProjectResult,
+  FxError,
+  Inputs,
+  InputsWithProjectPath,
+  Platform,
+  Result,
+  err,
+  ok,
+} from "@microsoft/teamsfx-api";
 import { DotenvParseOutput } from "dotenv";
 import fs from "fs-extra";
+import { glob } from "glob";
 import * as jsonschema from "jsonschema";
 import { camelCase, merge } from "lodash";
 import { EOL } from "os";
 import * as path from "path";
 import * as uuid from "uuid";
 import * as xml2js from "xml2js";
-
-import { hooks } from "@feathersjs/hooks/lib";
-import {
-  Colors,
-  Context,
-  CreateProjectResult,
-  err,
-  FxError,
-  Inputs,
-  InputsWithProjectPath,
-  ok,
-  Platform,
-  Result,
-} from "@microsoft/teamsfx-api";
-import { glob } from "glob";
 import { getLocalizedString } from "../../common/localizeUtils";
 import { TelemetryEvent, TelemetryProperty } from "../../common/telemetry";
 import { getResourceGroupInPortal } from "../../common/tools";
@@ -32,10 +31,10 @@ import { ObjectIsUndefinedError } from "../../core/error";
 import { ErrorContextMW, globalVars } from "../../core/globalVars";
 import { ResourceGroupConflictError, SelectSubscriptionError } from "../../error/azure";
 import {
-  assembleError,
   InputValidationError,
   MissingEnvironmentVariablesError,
   MissingRequiredInputError,
+  assembleError,
 } from "../../error/common";
 import { LifeCycleUndefinedError } from "../../error/yml";
 import {
@@ -61,10 +60,10 @@ import { SPFxGenerator } from "../generator/spfx/spfxGenerator";
 import { convertToLangKey } from "../generator/utils";
 import { ActionContext, ActionExecutionMW } from "../middleware/actionExecutionMW";
 import { provisionUtils } from "../provisionUtils";
+import { ResourceGroupInfo, resourceGroupHelper } from "../utils/ResourceGroupHelper";
 import { envUtil } from "../utils/envUtil";
 import { metadataUtil } from "../utils/metadataUtil";
 import { pathUtils } from "../utils/pathUtils";
-import { resourceGroupHelper, ResourceGroupInfo } from "../utils/ResourceGroupHelper";
 import { settingsUtil } from "../utils/settingsUtil";
 import { SummaryReporter } from "./summary";
 
@@ -142,8 +141,8 @@ const AzureDeployActions = [
 const needTenantCheckActions = ["botAadApp/create", "aadApp/create", "botFramework/create"];
 
 class Coordinator {
-  @ErrorContextMW({ component: "Coordinator" })
   @hooks([
+    ErrorContextMW({ component: "Coordinator" }),
     ActionExecutionMW({
       enableTelemetry: true,
       telemetryEventName: TelemetryEvent.CreateProject,
@@ -279,7 +278,6 @@ class Coordinator {
     return ok({ projectPath: projectPath, warnings });
   }
 
-  @ErrorContextMW({ component: "Coordinator" })
   async ensureTeamsFxInCsproj(projectPath: string): Promise<Result<undefined, FxError>> {
     const list = await fs.readdir(projectPath);
     const csprojFiles = list.filter((fileName) => fileName.endsWith(".csproj"));
@@ -315,7 +313,6 @@ class Coordinator {
     return ok(undefined);
   }
 
-  @ErrorContextMW({ component: "Coordinator" })
   async ensureTrackingId(
     projectPath: string,
     trackingId: string | undefined = undefined
@@ -330,7 +327,7 @@ class Coordinator {
     return ok(settings.trackingId);
   }
 
-  @ErrorContextMW({ component: "Coordinator" })
+  @hooks([ErrorContextMW({ component: "Coordinator" })])
   async preProvisionForVS(
     ctx: DriverContext,
     inputs: InputsWithProjectPath
@@ -399,7 +396,7 @@ class Coordinator {
     return ok(res);
   }
 
-  @ErrorContextMW({ component: "Coordinator" })
+  @hooks([ErrorContextMW({ component: "Coordinator" })])
   async preCheckYmlAndEnvForVS(
     ctx: DriverContext,
     inputs: InputsWithProjectPath
@@ -426,18 +423,10 @@ class Coordinator {
     return ok(undefined);
   }
 
-  @ErrorContextMW({ component: "Coordinator" })
-  @hooks([
-    ActionExecutionMW({
-      enableTelemetry: true,
-      telemetryEventName: TelemetryEvent.Provision,
-      telemetryComponentName: "coordinator",
-    }),
-  ])
+  @hooks([ErrorContextMW({ component: "Coordinator" })])
   async provision(
     ctx: DriverContext,
-    inputs: InputsWithProjectPath,
-    actionContext?: ActionContext
+    inputs: InputsWithProjectPath
   ): Promise<Result<DotenvParseOutput, FxError>> {
     const output: DotenvParseOutput = {};
     const folderName = path.parse(ctx.projectPath).name;
@@ -708,7 +697,6 @@ class Coordinator {
     return ok(output);
   }
 
-  @ErrorContextMW({ component: "Coordinator" })
   convertExecuteResult(
     execRes: Result<ExecutionOutput, ExecutionError>,
     templatePath: string
@@ -742,18 +730,10 @@ class Coordinator {
     return [output, error];
   }
 
-  @ErrorContextMW({ component: "Coordinator" })
-  @hooks([
-    ActionExecutionMW({
-      enableTelemetry: true,
-      telemetryEventName: TelemetryEvent.Deploy,
-      telemetryComponentName: "coordinator",
-    }),
-  ])
+  @hooks([ErrorContextMW({ component: "Coordinator" })])
   async deploy(
     ctx: DriverContext,
-    inputs: InputsWithProjectPath,
-    actionContext?: ActionContext
+    inputs: InputsWithProjectPath
   ): Promise<Result<DotenvParseOutput, FxError>> {
     const output: DotenvParseOutput = {};
     const templatePath =
@@ -823,18 +803,10 @@ class Coordinator {
     return ok(output);
   }
 
-  @ErrorContextMW({ component: "Coordinator" })
-  @hooks([
-    ActionExecutionMW({
-      enableTelemetry: true,
-      telemetryEventName: "publish",
-      telemetryComponentName: "coordinator",
-    }),
-  ])
+  @hooks([ErrorContextMW({ component: "Coordinator" })])
   async publish(
     ctx: DriverContext,
-    inputs: InputsWithProjectPath,
-    actionContext?: ActionContext
+    inputs: InputsWithProjectPath
   ): Promise<Result<DotenvParseOutput, FxError>> {
     const output: DotenvParseOutput = {};
     const templatePath = pathUtils.getYmlFilePath(ctx.projectPath, inputs.env);
@@ -892,19 +864,10 @@ class Coordinator {
     return ok(output);
   }
 
-  @ErrorContextMW({ component: "Coordinator" })
-  @hooks([
-    ActionExecutionMW({
-      enableTelemetry: true,
-      telemetryEventName: TelemetryEvent.PublishInDeveloperPortal,
-      telemetryComponentName: "coordinator",
-      errorSource: CoordinatorSource,
-    }),
-  ])
+  @hooks([ErrorContextMW({ component: "Coordinator" })])
   async publishInDeveloperPortal(
     ctx: Context,
-    inputs: InputsWithProjectPath,
-    actionContext?: ActionContext
+    inputs: InputsWithProjectPath
   ): Promise<Result<undefined, FxError>> {
     // update teams app
     if (!ctx.tokenProvider) {
