+++ packages/fx-core/src/component/feature/collaboration.ts
@@ -32,9 +32,10 @@ export class AadCollaboration {
   constructor(m365TokenProvider: M365TokenProvider) {
     this.aadAppClient = new AadAppClient(m365TokenProvider);
   }
-
-  @ErrorContextMW({ source: "Graph" })
-  @hooks([addStartAndEndTelemetry(EventName.grantPermission, componentNameAad)])
+  @hooks([
+    ErrorContextMW({ source: "Graph", component: "AadCollaboration" }),
+    addStartAndEndTelemetry(EventName.grantPermission, componentNameAad),
+  ])
   public async grantPermission(
     ctx: Context,
     objectId: string,
@@ -56,8 +57,10 @@ export class AadCollaboration {
       return err(this.handleError(error, ctx, objectId));
     }
   }
-  @ErrorContextMW({ source: "Graph" })
-  @hooks([addStartAndEndTelemetry(EventName.listCollaborator, componentNameAad)])
+  @hooks([
+    ErrorContextMW({ source: "Graph", component: "AadCollaboration" }),
+    addStartAndEndTelemetry(EventName.listCollaborator, componentNameAad),
+  ])
   public async listCollaborator(
     ctx: Context,
     objectId: string
@@ -69,8 +72,10 @@ export class AadCollaboration {
       return err(this.handleError(error, ctx, objectId));
     }
   }
-  @ErrorContextMW({ source: "Graph" })
-  @hooks([addStartAndEndTelemetry(EventName.checkPermission, componentNameAad)])
+  @hooks([
+    ErrorContextMW({ source: "Graph", component: "AadCollaboration" }),
+    addStartAndEndTelemetry(EventName.checkPermission, componentNameAad),
+  ])
   public async checkPermission(
     ctx: Context,
     objectId: string,
@@ -121,8 +126,10 @@ export class TeamsCollaboration {
     this.tokenProvider = m365TokenProvider;
     TelemetryUtils.init(ctx);
   }
-  @ErrorContextMW({ source: "Teams" })
-  @hooks([addStartAndEndTelemetry(EventName.grantPermission, componentNameTeams)])
+  @hooks([
+    ErrorContextMW({ source: "Teams", component: "TeamsCollaboration" }),
+    addStartAndEndTelemetry(EventName.grantPermission, componentNameTeams),
+  ])
   public async grantPermission(
     ctx: Context,
     teamsAppId: string,
@@ -148,8 +155,10 @@ export class TeamsCollaboration {
       return err(this.handleError(error, ctx, teamsAppId));
     }
   }
-  @ErrorContextMW({ source: "Teams" })
-  @hooks([addStartAndEndTelemetry(EventName.listCollaborator, componentNameTeams)])
+  @hooks([
+    ErrorContextMW({ source: "Teams", component: "TeamsCollaboration" }),
+    addStartAndEndTelemetry(EventName.listCollaborator, componentNameTeams),
+  ])
   public async listCollaborator(
     ctx: Context,
     teamsAppId: string
@@ -183,8 +192,10 @@ export class TeamsCollaboration {
       return err(this.handleError(error, ctx, teamsAppId));
     }
   }
-  @ErrorContextMW({ source: "Teams" })
-  @hooks([addStartAndEndTelemetry(EventName.checkPermission, componentNameTeams)])
+  @hooks([
+    ErrorContextMW({ source: "Teams", component: "TeamsCollaboration" }),
+    addStartAndEndTelemetry(EventName.checkPermission, componentNameTeams),
+  ])
   public async checkPermission(
     ctx: Context,
     teamsAppId: string,
