+++ packages/fx-core/src/component/resource/botService/appStudio/appStudioClient.ts
@@ -26,6 +26,7 @@ import { Messages } from "../messages";
 import { RetryHandler } from "../retryHandler";
 import { CommonStrings, ConfigNames } from "../strings";
 import { ErrorContextMW } from "../../../../core/globalVars";
+import { hooks } from "@feathersjs/hooks";
 
 function handleBotFrameworkError(e: any, apiName: string): void | undefined {
   if (e.response?.status === HttpStatusCode.NOTFOUND) {
@@ -73,7 +74,7 @@ export class AppStudioClient {
   public static setRegion(region: string) {
     AppStudioClient.baseUrl = region;
   }
-  @ErrorContextMW({ source: "Teams" })
+  @hooks([ErrorContextMW({ source: "Teams", component: "AppStudioClient" })])
   public static async getBotRegistration(
     token: string,
     botId: string
@@ -97,7 +98,7 @@ export class AppStudioClient {
       handleBotFrameworkError(e, APP_STUDIO_API_NAMES.GET_BOT);
     }
   }
-  @ErrorContextMW({ source: "Teams" })
+  @hooks([ErrorContextMW({ source: "Teams", component: "AppStudioClient" })])
   public static async createBotRegistration(
     token: string,
     registration: IBotRegistration,
@@ -150,7 +151,7 @@ export class AppStudioClient {
 
     return;
   }
-  @ErrorContextMW({ source: "Teams" })
+  @hooks([ErrorContextMW({ source: "Teams", component: "AppStudioClient" })])
   public static async updateBotRegistration(
     token: string,
     botReg: IBotRegistration
