+++ packages/fx-core/src/core/FxCore.ts
@@ -3,7 +3,11 @@
 
 import { hooks } from "@feathersjs/hooks";
 import {
+  AdaptiveFolderName,
   ApiOperation,
+  AppPackageFolderName,
+  BuildFolderName,
+  Context,
   CoreCallbackEvent,
   CreateProjectResult,
   CryptoProvider,
@@ -14,6 +18,7 @@ import {
   InputsWithProjectPath,
   ok,
   OpenAIPluginManifest,
+  Platform,
   QTreeNode,
   Result,
   Stage,
@@ -21,39 +26,93 @@ import {
 } from "@microsoft/teamsfx-api";
 import { DotenvParseOutput } from "dotenv";
 import fs from "fs-extra";
+import * as os from "os";
 import * as path from "path";
 import "reflect-metadata";
+import { Container } from "typedi";
+import { pathToFileURL } from "url";
 import { parse } from "yaml";
+import { VSCodeExtensionCommand } from "../common/constants";
+import { getLocalizedString } from "../common/localizeUtils";
+import { LaunchHelper } from "../common/m365/launchHelper";
 import { ListCollaboratorResult, PermissionsResult } from "../common/permissionInterface";
+import { isValidProjectV2, isValidProjectV3 } from "../common/projectSettingsHelper";
+import { SpecParser } from "../common/spec-parser/specParser";
+import { SpecParserError } from "../common/spec-parser/specParserError";
+import { VersionSource, VersionState } from "../common/versionMetadata";
 import { ILifecycle, LifecycleName } from "../component/configManager/interface";
 import { YamlParser } from "../component/configManager/parser";
+import {
+  AadConstants,
+  SingleSignOnOptionItem,
+  SPFxQuestionNames,
+  ViewAadAppHelpLinkV5,
+} from "../component/constants";
+import { coordinator } from "../component/coordinator";
+import { UpdateAadAppArgs } from "../component/driver/aad/interface/updateAadAppArgs";
+import { UpdateAadAppDriver } from "../component/driver/aad/update";
+import { buildAadManifest } from "../component/driver/aad/utility/buildAadManifest";
+import { AddWebPartDriver } from "../component/driver/add/addWebPart";
+import { AddWebPartArgs } from "../component/driver/add/interface/AddWebPartArgs";
 import "../component/driver/index";
 import { DriverContext } from "../component/driver/interface/commonArgs";
 import "../component/driver/script/scriptDriver";
-import { ErrorResult } from "../component/generator/copilotPlugin/helper";
-import { EnvLoaderMW } from "../component/middleware/envMW";
+import { updateManifestV3 } from "../component/driver/teamsApp/appStudio";
+import { CreateAppPackageDriver } from "../component/driver/teamsApp/createAppPackage";
+import { CreateAppPackageArgs } from "../component/driver/teamsApp/interfaces/CreateAppPackageArgs";
+import { ValidateAppPackageArgs } from "../component/driver/teamsApp/interfaces/ValidateAppPackageArgs";
+import { ValidateManifestArgs } from "../component/driver/teamsApp/interfaces/ValidateManifestArgs";
+import { manifestUtils } from "../component/driver/teamsApp/utils/ManifestUtils";
+import {
+  containsUnsupportedFeature,
+  getFeaturesFromAppDefinition,
+} from "../component/driver/teamsApp/utils/utils";
+import { ValidateManifestDriver } from "../component/driver/teamsApp/validate";
+import { ValidateAppPackageDriver } from "../component/driver/teamsApp/validateAppPackage";
+import { SSO } from "../component/feature/sso";
+import {
+  convertSpecParserErrorToFxError,
+  ErrorResult,
+  listOperations,
+  OpenAIPluginManifestHelper,
+} from "../component/generator/copilotPlugin/helper";
+import { EnvLoaderMW, EnvWriterMW } from "../component/middleware/envMW";
+import { QuestionMW } from "../component/middleware/questionMW";
+import { createContextV3, createDriverContext } from "../component/utils";
 import { envUtil } from "../component/utils/envUtil";
 import { metadataUtil } from "../component/utils/metadataUtil";
 import { pathUtils } from "../component/utils/pathUtils";
 import { settingsUtil } from "../component/utils/settingsUtil";
-import { createProjectCliHelpNode } from "../question/create";
+import { assembleError, FileNotFoundError, InvalidProjectError } from "../error/common";
+import { NoNeedUpgradeError } from "../error/upgrade";
+import { YamlFieldMissingError } from "../error/yml";
+import { ValidateTeamsAppInputs } from "../question";
+import { createProjectCliHelpNode, ScratchOptions, SPFxVersionOptionIds } from "../question/create";
+import { HubTypes, isAadMainifestContainsPlaceholder } from "../question/other";
+import { QuestionNames } from "../question/questionNames";
 import { CallbackRegistry } from "./callback";
+import { checkPermission, grantPermission, listCollaborator } from "./collaborator";
 import { LocalCrypto } from "./crypto";
 import { environmentManager } from "./environment";
 import { InvalidInputError } from "./error";
-import { FxCoreV3Implement } from "./FxCoreImplementV3";
-import { ErrorContextMW, setTools, TOOLS } from "./globalVars";
+import { ErrorContextMW, setErrorContext, setTools, TOOLS } from "./globalVars";
+import { ConcurrentLockerMW } from "./middleware/concurrentLocker";
+import { ContextInjectorMW } from "./middleware/contextInjector";
 import { ErrorHandlerMW } from "./middleware/errorHandler";
-import { PreProvisionResForVS, VersionCheckRes } from "./types";
+import { checkActiveResourcePlugins, ProjectMigratorMWV3 } from "./middleware/projectMigratorV3";
+import {
+  getProjectVersionFromPath,
+  getTrackingIdFromPath,
+  getVersionState,
+} from "./middleware/utils/v3MigrationUtils";
+import { CoreTelemetryEvent, CoreTelemetryProperty } from "./telemetry";
+import { CoreHookContext, PreProvisionResForVS, VersionCheckRes } from "./types";
 
 export type CoreCallbackFunc = (name: string, err?: FxError, data?: any) => void | Promise<void>;
 
 export class FxCore {
-  v3Implement: FxCoreV3Implement;
-
   constructor(tools: Tools) {
     setTools(tools);
-    this.v3Implement = new FxCoreV3Implement(tools);
   }
 
   /**
@@ -64,39 +123,111 @@ export class FxCore {
     return CallbackRegistry.set(event, callback);
   }
 
-  /**
-   * lifecycle command: create new project
-   */
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "createProject" })])
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "createProject", reset: true }),
+    ErrorHandlerMW,
+    QuestionMW("createProject"),
+  ])
   async createProject(inputs: Inputs): Promise<Result<CreateProjectResult, FxError>> {
-    return this.v3Implement.dispatch(this.createProject, inputs);
+    const context = createContextV3();
+    inputs[QuestionNames.Scratch] = ScratchOptions.yes().id;
+    if (inputs.teamsAppFromTdp) {
+      // should never happen as we do same check on Developer Portal.
+      if (containsUnsupportedFeature(inputs.teamsAppFromTdp)) {
+        return err(InvalidInputError("Teams app contains unsupported features"));
+      } else {
+        context.telemetryReporter.sendTelemetryEvent(CoreTelemetryEvent.CreateFromTdpStart, {
+          [CoreTelemetryProperty.TdpTeamsAppFeatures]: getFeaturesFromAppDefinition(
+            inputs.teamsAppFromTdp
+          ).join(","),
+          [CoreTelemetryProperty.TdpTeamsAppId]: inputs.teamsAppFromTdp.teamsAppId,
+        });
+      }
+    }
+    const res = await coordinator.create(context, inputs);
+    inputs.projectPath = context.projectPath;
+    return res;
   }
 
   /**
    * lifecycle command: create new sample project
    */
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "createSampleProject" })])
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "createSampleProject", reset: true }),
+    ErrorHandlerMW,
+    QuestionMW("createSampleProject"),
+  ])
   async createSampleProject(inputs: Inputs): Promise<Result<CreateProjectResult, FxError>> {
-    return this.v3Implement.dispatch(this.createSampleProject, inputs);
+    const context = createContextV3();
+    inputs[QuestionNames.Scratch] = ScratchOptions.no().id;
+    const res = await coordinator.create(context, inputs);
+    inputs.projectPath = context.projectPath;
+    return res;
   }
-
   /**
    * lifecycle commands: provision
    */
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "provision" })])
-  async provisionResources(inputs: Inputs): Promise<Result<undefined, FxError>> {
-    return this.v3Implement.dispatch(this.provisionResources, inputs);
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "provision", reset: true }),
+    ErrorHandlerMW,
+    ProjectMigratorMWV3,
+    EnvLoaderMW(false),
+    ConcurrentLockerMW,
+    ContextInjectorMW,
+    EnvWriterMW,
+  ])
+  async provisionResources(
+    inputs: Inputs,
+    ctx?: CoreHookContext
+  ): Promise<Result<undefined, FxError>> {
+    inputs.stage = Stage.provision;
+    const context = createDriverContext(inputs);
+    try {
+      const res = await coordinator.provision(context, inputs as InputsWithProjectPath);
+      if (res.isOk()) {
+        ctx!.envVars = res.value;
+        return ok(undefined);
+      } else {
+        // for partial success scenario, output is set in inputs object
+        ctx!.envVars = inputs.envVars;
+        return err(res.error);
+      }
+    } finally {
+      //reset subscription
+      try {
+        await TOOLS.tokenProvider.azureAccountProvider.setSubscription("");
+      } catch (e) {}
+    }
   }
-
   /**
    * lifecycle commands: deploy
    */
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "deploy" })])
-  async deployArtifacts(inputs: Inputs): Promise<Result<undefined, FxError>> {
-    return this.v3Implement.dispatch(this.deployArtifacts, inputs);
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "deploy", reset: true }),
+    ErrorHandlerMW,
+    ProjectMigratorMWV3,
+    EnvLoaderMW(false),
+    ConcurrentLockerMW,
+    ContextInjectorMW,
+    EnvWriterMW,
+  ])
+  async deployArtifacts(
+    inputs: Inputs,
+    ctx?: CoreHookContext
+  ): Promise<Result<undefined, FxError>> {
+    inputs.stage = Stage.deploy;
+    const context = createDriverContext(inputs);
+    const res = await coordinator.deploy(context, inputs as InputsWithProjectPath);
+    if (res.isOk()) {
+      ctx!.envVars = res.value;
+      return ok(undefined);
+    } else {
+      // for partial success scenario, output is set in inputs object
+      ctx!.envVars = inputs.envVars;
+      return err(res.error);
+    }
   }
-
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "localDebug" })])
+  @hooks([ErrorContextMW({ component: "FxCore", stage: "localDebug", reset: true })])
   async localDebug(inputs: Inputs): Promise<Result<undefined, FxError>> {
     inputs.env = environmentManager.getLocalEnvName();
     return this.provisionResources(inputs);
@@ -105,91 +236,326 @@ export class FxCore {
   /**
    * none lifecycle command, v3 only
    */
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "deployAadManifest" })])
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "deployAadManifest", reset: true }),
+    ErrorHandlerMW,
+    ProjectMigratorMWV3,
+    QuestionMW("deployAadManifest"),
+    EnvLoaderMW(true, true),
+    ConcurrentLockerMW,
+    ContextInjectorMW,
+  ])
   async deployAadManifest(inputs: Inputs): Promise<Result<undefined, FxError>> {
-    return this.v3Implement.dispatch(this.deployAadManifest, inputs);
+    inputs.stage = Stage.deployAad;
+    const updateAadClient = Container.get<UpdateAadAppDriver>("aadApp/update");
+    // In V3, the aad.template.json exist at .fx folder, and output to root build folder.
+    const manifestTemplatePath: string = inputs[QuestionNames.AadAppManifestFilePath];
+    if (!(await fs.pathExists(manifestTemplatePath))) {
+      return err(new FileNotFoundError("deployAadManifest", manifestTemplatePath));
+    }
+    let manifestOutputPath: string = manifestTemplatePath;
+    if (inputs.env && (await isAadMainifestContainsPlaceholder(inputs))) {
+      await fs.ensureDir(path.join(inputs.projectPath!, "build"));
+      manifestOutputPath = path.join(
+        inputs.projectPath!,
+        "build",
+        `aad.manifest.${inputs.env}.json`
+      );
+    }
+    const inputArgs: UpdateAadAppArgs = {
+      manifestPath: manifestTemplatePath,
+      outputFilePath: manifestOutputPath,
+    };
+    const Context: DriverContext = createDriverContext(inputs);
+    setErrorContext({ component: "aadAppUpdate" });
+    const res = await updateAadClient.run(inputArgs, Context);
+    if (res.isErr()) {
+      return err(res.error);
+    }
+    if (Context.platform === Platform.CLI) {
+      const msg = getLocalizedString("core.deploy.aadManifestOnCLISuccessNotice");
+      Context.ui!.showMessage("info", msg, false);
+    } else {
+      const msg = getLocalizedString("core.deploy.aadManifestSuccessNotice");
+      Context.ui!.showMessage(
+        "info",
+        msg,
+        false,
+        getLocalizedString("core.deploy.aadManifestLearnMore")
+      ).then((result) => {
+        const userSelected = result.isOk() ? result.value : undefined;
+        if (userSelected === getLocalizedString("core.deploy.aadManifestLearnMore")) {
+          Context.ui!.openUrl(ViewAadAppHelpLinkV5);
+        }
+      });
+    }
+    return ok(undefined);
   }
-
   /**
    * none lifecycle command, v3 only
    */
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "addWebpart" })])
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "addWebpart", reset: true }),
+    ErrorHandlerMW,
+    QuestionMW("addWebpart"),
+    ProjectMigratorMWV3,
+    ConcurrentLockerMW,
+  ])
   async addWebpart(inputs: Inputs): Promise<Result<undefined, FxError>> {
-    return this.v3Implement.dispatch(this.addWebpart, inputs);
+    setErrorContext({ component: "spfxAdd", method: "run" });
+    const driver: AddWebPartDriver = Container.get<AddWebPartDriver>("spfx/add");
+    const args: AddWebPartArgs = {
+      manifestPath: inputs[SPFxQuestionNames.ManifestPath],
+      localManifestPath: inputs[SPFxQuestionNames.LocalManifestPath],
+      spfxFolder: inputs[SPFxQuestionNames.SPFxFolder],
+      webpartName: inputs[SPFxQuestionNames.WebPartName],
+      spfxPackage: SPFxVersionOptionIds.installLocally,
+    };
+    const Context: DriverContext = createDriverContext(inputs);
+    const res = await driver.run(args, Context);
+    if (res.isErr()) {
+      return err(res.error);
+    }
+    return ok(undefined);
   }
-
   /**
    * lifecycle command: publish
    */
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "publish" })])
-  async publishApplication(inputs: Inputs): Promise<Result<undefined, FxError>> {
-    return this.v3Implement.dispatch(this.publishApplication, inputs);
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "publish", reset: true }),
+    ErrorHandlerMW,
+    ProjectMigratorMWV3,
+    EnvLoaderMW(false),
+    ConcurrentLockerMW,
+    ContextInjectorMW,
+    EnvWriterMW,
+  ])
+  async publishApplication(
+    inputs: Inputs,
+    ctx?: CoreHookContext
+  ): Promise<Result<undefined, FxError>> {
+    inputs.stage = Stage.publish;
+    const context = createDriverContext(inputs);
+    const res = await coordinator.publish(context, inputs as InputsWithProjectPath);
+    if (res.isOk()) {
+      ctx!.envVars = res.value;
+      return ok(undefined);
+    } else {
+      // for partial success scenario, output is set in inputs object
+      ctx!.envVars = inputs.envVars;
+      return err(res.error);
+    }
   }
-
   /**
    * most commands will be deprecated in V3
    */
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "executeUserTask" })])
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "executeUserTask", reset: true }),
+    ErrorHandlerMW,
+    ProjectMigratorMWV3,
+    EnvLoaderMW(false),
+    ConcurrentLockerMW,
+  ])
   async executeUserTask(func: Func, inputs: Inputs): Promise<Result<any, FxError>> {
-    return await this.v3Implement.dispatchUserTask(this.executeUserTask, func, inputs);
+    let res: Result<any, FxError> = ok(undefined);
+    const context = createDriverContext(inputs);
+    if (func.method === "addSso") {
+      // used in v3 only in VS
+      inputs.stage = Stage.addFeature;
+      inputs[QuestionNames.Features] = SingleSignOnOptionItem.id;
+      const component = Container.get<SSO>("sso");
+      setErrorContext({ component: "sso", method: "add" });
+      res = await component.add(context as unknown as Context, inputs as InputsWithProjectPath);
+    }
+    return res;
   }
-
   /**
    * v3 only none lifecycle command
    */
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "buildAadManifest" })])
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "buildAadManifest", reset: true }),
+    ErrorHandlerMW,
+    ProjectMigratorMWV3,
+    EnvLoaderMW(false),
+    ConcurrentLockerMW,
+  ])
   async buildAadManifest(inputs: Inputs): Promise<Result<undefined, FxError>> {
-    return this.v3Implement.dispatch(this.buildAadManifest, inputs);
-  }
-
-  /**
-   * v3 only none lifecycle command
-   */
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "deployTeamsManifest" })])
-  async deployTeamsManifest(inputs: Inputs): Promise<Result<undefined, FxError>> {
-    return this.v3Implement.dispatch(this.deployTeamsManifest, inputs);
-  }
-
-  /**
-   * v3 only none lifecycle command
-   */
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "validateApplication" })])
-  async validateApplication(inputs: Inputs): Promise<Result<any, FxError>> {
-    return this.v3Implement.dispatch(this.validateApplication, inputs);
+    const manifestTemplatePath: string = inputs.AAD_MANIFEST_FILE
+      ? inputs.AAD_MANIFEST_FILE
+      : path.join(inputs.projectPath!, AadConstants.DefaultTemplateFileName);
+    if (!(await fs.pathExists(manifestTemplatePath))) {
+      return err(new FileNotFoundError("buildAadManifest", manifestTemplatePath));
+    }
+    await fs.ensureDir(path.join(inputs.projectPath!, "build"));
+    const manifestOutputPath: string = path.join(
+      inputs.projectPath!,
+      "build",
+      `aad.${inputs.env}.json`
+    );
+    const Context: DriverContext = createDriverContext(inputs);
+    await buildAadManifest(Context, manifestTemplatePath, manifestOutputPath);
+    return ok(undefined);
   }
   /**
    * v3 only none lifecycle command
    */
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "validateManifest" })])
-  async validateManifest(inputs: Inputs): Promise<Result<any, FxError>> {
-    return this.v3Implement.dispatch(this.validateManifest, inputs);
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "deployTeamsManifest", reset: true }),
+    ErrorHandlerMW,
+    ProjectMigratorMWV3,
+    QuestionMW("selectTeamsAppManifest"),
+    EnvLoaderMW(true),
+    ConcurrentLockerMW,
+    ContextInjectorMW,
+    EnvWriterMW,
+  ])
+  async deployTeamsManifest(
+    inputs: Inputs,
+    ctx?: CoreHookContext
+  ): Promise<Result<undefined, FxError>> {
+    inputs.manifestTemplatePath = inputs[QuestionNames.TeamsAppManifestFilePath] as string;
+    const context = createContextV3();
+    const res = await updateManifestV3(context, inputs as InputsWithProjectPath);
+    if (res.isOk()) {
+      ctx!.envVars = envUtil.map2object(res.value);
+      return ok(undefined);
+    }
+    return err(res.error);
   }
   /**
    * v3 only none lifecycle command
    */
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "validateAppPackage" })])
-  async validateAppPackage(inputs: Inputs): Promise<Result<any, FxError>> {
-    return this.v3Implement.dispatch(this.validateAppPackage, inputs);
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "validateApplication", reset: true }),
+    QuestionMW("validateTeamsApp"),
+  ])
+  async validateApplication(inputs: ValidateTeamsAppInputs): Promise<Result<any, FxError>> {
+    if (inputs["manifest-path"]) {
+      return await this.validateManifest(inputs);
+    } else {
+      return await this.validateAppPackage(inputs);
+    }
   }
   /**
    * v3 only none lifecycle command
    */
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "createAppPackage" })])
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "validateManifest", reset: true }),
+    ErrorHandlerMW,
+    EnvLoaderMW(true),
+    ConcurrentLockerMW,
+  ])
+  async validateManifest(inputs: ValidateTeamsAppInputs): Promise<Result<any, FxError>> {
+    inputs.stage = Stage.validateApplication;
+    const context: DriverContext = createDriverContext(inputs);
+    const teamsAppManifestFilePath = inputs["manifest-path"] as string;
+    const args: ValidateManifestArgs = {
+      manifestPath: teamsAppManifestFilePath,
+      showMessage: inputs?.showMessage != undefined ? inputs.showMessage : true,
+    };
+    const driver: ValidateManifestDriver = Container.get("teamsApp/validateManifest");
+    const result = await driver.run(args, context);
+    return result;
+  }
+  /**
+   * v3 only none lifecycle command
+   */
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "validateAppPackage", reset: true }),
+    ErrorHandlerMW,
+    ConcurrentLockerMW,
+  ])
+  async validateAppPackage(inputs: ValidateTeamsAppInputs): Promise<Result<any, FxError>> {
+    inputs.stage = Stage.validateApplication;
+    const context: DriverContext = createDriverContext(inputs);
+    const teamsAppPackageFilePath = inputs["app-package-file-path"] as string;
+    const args: ValidateAppPackageArgs = {
+      appPackagePath: teamsAppPackageFilePath,
+      showMessage: true,
+    };
+    const driver: ValidateAppPackageDriver = Container.get("teamsApp/validateAppPackage");
+    return await driver.run(args, context);
+  }
+  /**
+   * v3 only none lifecycle command
+   */
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "createAppPackage", reset: true }),
+    ErrorHandlerMW,
+    QuestionMW("selectTeamsAppManifest"),
+    EnvLoaderMW(true),
+    ConcurrentLockerMW,
+  ])
   async createAppPackage(inputs: Inputs): Promise<Result<any, FxError>> {
-    return this.v3Implement.dispatch(this.createAppPackage, inputs);
-  }
+    inputs.stage = Stage.createAppPackage;
 
+    const context: DriverContext = createDriverContext(inputs);
+
+    const teamsAppManifestFilePath = inputs?.[QuestionNames.TeamsAppManifestFilePath] as string;
+
+    const driver: CreateAppPackageDriver = Container.get("teamsApp/zipAppPackage");
+    const args: CreateAppPackageArgs = {
+      manifestPath: teamsAppManifestFilePath,
+      outputZipPath:
+        inputs[QuestionNames.OutputZipPathParamName] ??
+        `${inputs.projectPath}/${AppPackageFolderName}/${BuildFolderName}/appPackage.${process.env.TEAMSFX_ENV}.zip`,
+      outputJsonPath:
+        inputs[QuestionNames.OutputManifestParamName] ??
+        `${inputs.projectPath}/${AppPackageFolderName}/${BuildFolderName}/manifest.${process.env.TEAMSFX_ENV}.json`,
+    };
+    const result = await driver.run(args, context);
+    if (context.platform === Platform.VSCode) {
+      if (result.isOk()) {
+        const isWindows = process.platform === "win32";
+        let zipFileName = args.outputZipPath;
+        if (!path.isAbsolute(zipFileName)) {
+          zipFileName = path.join(context.projectPath, zipFileName);
+        }
+        let builtSuccess = getLocalizedString(
+          "plugins.appstudio.buildSucceedNotice.fallback",
+          zipFileName
+        );
+        if (isWindows) {
+          const folderLink = pathToFileURL(path.dirname(zipFileName));
+          const appPackageLink = `${VSCodeExtensionCommand.openFolder}?%5B%22${folderLink}%22%5D`;
+          builtSuccess = getLocalizedString("plugins.appstudio.buildSucceedNotice", appPackageLink);
+        }
+        context.ui?.showMessage("info", builtSuccess, false);
+      }
+    }
+    return result;
+  }
   /**
    * get url to preview the app, may prompt to select env, hub and Teams manifest
    * v3 only none lifecycle command
    * @param {Inputs} inputs
    * @returns the url to preview the app
    */
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "preview" })])
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "preview", reset: true }),
+    ErrorHandlerMW,
+    QuestionMW("previewWithTeamsAppManifest"),
+    EnvLoaderMW(false),
+    ConcurrentLockerMW,
+  ])
   async previewWithManifest(inputs: Inputs): Promise<Result<string, FxError>> {
-    return this.v3Implement.dispatch(this.previewWithManifest, inputs);
-  }
+    inputs.stage = Stage.previewWithManifest;
 
+    const hub = inputs[QuestionNames.M365Host] as HubTypes;
+    const manifestFilePath = inputs[QuestionNames.TeamsAppManifestFilePath] as string;
+
+    const manifestRes = await manifestUtils.getManifestV3(manifestFilePath, false);
+    if (manifestRes.isErr()) {
+      return err(manifestRes.error);
+    }
+
+    const teamsAppId = manifestRes.value.id;
+    const capabilities = manifestUtils.getCapabilities(manifestRes.value);
+
+    const launchHelper = new LaunchHelper(TOOLS.tokenProvider.m365TokenProvider, TOOLS.logProvider);
+    const result = await launchHelper.getLaunchUrl(hub, teamsAppId, capabilities);
+    return result;
+  }
   /**
    * Warning: this API only works for CLI_HELP, it has no business with interactive run for CLI!
    */
@@ -208,17 +574,31 @@ export class FxCore {
   /**
    * get all dot envs
    */
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "getDotEnvs" })])
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "getDotEnvs", reset: true }),
+    ErrorHandlerMW,
+  ])
   async getDotEnvs(
     inputs: InputsWithProjectPath
   ): Promise<Result<{ [name: string]: DotenvParseOutput }, FxError>> {
-    return this.v3Implement.dispatch(this.getDotEnvs, inputs);
+    const envListRes = await envUtil.listEnv(inputs.projectPath);
+    if (envListRes.isErr()) {
+      return err(envListRes.error);
+    }
+    const res: { [name: string]: DotenvParseOutput } = {};
+    for (const env of envListRes.value) {
+      const envRes = await envUtil.readEnv(inputs.projectPath, env, false, false);
+      if (envRes.isErr()) {
+        return err(envRes.error);
+      }
+      res[env] = envRes.value as DotenvParseOutput;
+    }
+    return ok(res);
   }
-
   /**
    * given projectPath and filePath, return whether the filePath is a env file
    */
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "isEnvFile" })])
+  @hooks([ErrorContextMW({ component: "FxCore", stage: "isEnvFile", reset: true })])
   async isEnvFile(projectPath: string, inputFile: string): Promise<Result<boolean, FxError>> {
     const inputFileName = path.basename(inputFile);
     const envName = envUtil.extractEnvNameFromFileName(inputFileName);
@@ -235,7 +615,7 @@ export class FxCore {
   /**
    * get projectId
    */
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "getProjectId" })])
+  @hooks([ErrorContextMW({ component: "FxCore", stage: "getProjectId", reset: true })])
   async getProjectId(projectPath: string): Promise<Result<string, FxError>> {
     const res = await this.getProjectMetadata(projectPath);
     if (res.isErr()) {
@@ -247,7 +627,7 @@ export class FxCore {
   /**
    * @description get projectId and version from yml
    */
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "getProjectMetadata" })])
+  @hooks([ErrorContextMW({ component: "FxCore", stage: "getProjectMetadata", reset: true })])
   async getProjectMetadata(
     projectPath: string
   ): Promise<Result<{ version?: string; projectId?: string }, FxError>> {
@@ -270,7 +650,7 @@ export class FxCore {
   /**
    * get Teams App Name from yml
    */
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "getTeamsAppName" })])
+  @hooks([ErrorContextMW({ component: "FxCore", stage: "getTeamsAppName", reset: true })])
   async getTeamsAppName(projectPath: string): Promise<Result<string, FxError>> {
     const ymlPath = pathUtils.getYmlFilePath(projectPath, "dev");
     const maybeProjectModel = await metadataUtil.parse(ymlPath, "dev");
@@ -295,7 +675,7 @@ export class FxCore {
   /**
    * get project info
    */
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "getProjectInfo" })])
+  @hooks([ErrorContextMW({ component: "FxCore", stage: "getProjectInfo", reset: true })])
   async getProjectInfo(
     projectPath: string,
     env: string
@@ -351,29 +731,71 @@ export class FxCore {
     return ok(res);
   }
 
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "grantPermission" })])
-  async grantPermission(inputs: Inputs): Promise<Result<PermissionsResult, FxError>> {
-    return this.v3Implement.dispatch(this.grantPermission, inputs);
-  }
-
-  /**
-   * none lifecycle command
-   */
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "checkPermission" })])
-  async checkPermission(inputs: Inputs): Promise<Result<PermissionsResult, FxError>> {
-    return this.v3Implement.dispatch(this.checkPermission, inputs);
-  }
-
-  /**
-   * none lifecycle command
-   */
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "listCollaborator" })])
-  async listCollaborator(inputs: Inputs): Promise<Result<ListCollaboratorResult, FxError>> {
-    return this.v3Implement.dispatch(this.listCollaborator, inputs);
-  }
-
   @hooks([
-    ErrorContextMW({ component: "FxCore", stage: "getSelectedEnv" }),
+    ErrorContextMW({ component: "FxCore", stage: "grantPermission", reset: true }),
+    ErrorHandlerMW,
+    ProjectMigratorMWV3,
+    QuestionMW("grantPermission"),
+    EnvLoaderMW(false, true),
+    ConcurrentLockerMW,
+    EnvWriterMW,
+  ])
+  async grantPermission(inputs: Inputs): Promise<Result<PermissionsResult, FxError>> {
+    inputs.stage = Stage.grantPermission;
+    const context = createContextV3();
+    setErrorContext({ component: "collaborator" });
+    const res = await grantPermission(
+      context,
+      inputs as InputsWithProjectPath,
+      TOOLS.tokenProvider
+    );
+    return res;
+  }
+  /**
+   * none lifecycle command
+   */
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "checkPermission", reset: true }),
+    ErrorHandlerMW,
+    ProjectMigratorMWV3,
+    EnvLoaderMW(false, true),
+    ConcurrentLockerMW,
+    EnvWriterMW,
+  ])
+  async checkPermission(inputs: Inputs): Promise<Result<PermissionsResult, FxError>> {
+    inputs.stage = Stage.checkPermission;
+    const context = createContextV3();
+    const res = await checkPermission(
+      context,
+      inputs as InputsWithProjectPath,
+      TOOLS.tokenProvider
+    );
+    return res;
+  }
+  /**
+   * none lifecycle command
+   */
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "listCollaborator", reset: true }),
+    ErrorHandlerMW,
+    ProjectMigratorMWV3,
+    QuestionMW("listCollaborator"),
+    EnvLoaderMW(false, true),
+    ConcurrentLockerMW,
+    EnvWriterMW,
+  ])
+  async listCollaborator(inputs: Inputs): Promise<Result<ListCollaboratorResult, FxError>> {
+    inputs.stage = Stage.listCollaborator;
+    const context = createContextV3();
+    const res = await listCollaborator(
+      context,
+      inputs as InputsWithProjectPath,
+      TOOLS.tokenProvider
+    );
+    return res;
+  }
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "getSelectedEnv", reset: true }),
     ErrorHandlerMW,
     EnvLoaderMW(false),
   ])
@@ -395,7 +817,7 @@ export class FxCore {
   /**
    * only for vs code extension
    */
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "encrypt" }), ErrorHandlerMW])
+  @hooks([ErrorContextMW({ component: "FxCore", stage: "encrypt", reset: true }), ErrorHandlerMW])
   async encrypt(plaintext: string, inputs: Inputs): Promise<Result<string, FxError>> {
     const res = await this.createLocalCrypto(inputs.projectPath!);
     if (res.isErr()) {
@@ -406,7 +828,7 @@ export class FxCore {
   /**
    * only for vs code extension
    */
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "decrypt" }), ErrorHandlerMW])
+  @hooks([ErrorContextMW({ component: "FxCore", stage: "decrypt", reset: true }), ErrorHandlerMW])
   async decrypt(ciphertext: string, inputs: Inputs): Promise<Result<string, FxError>> {
     const res = await this.createLocalCrypto(inputs.projectPath!);
     if (res.isErr()) {
@@ -415,22 +837,116 @@ export class FxCore {
     return res.value.decrypt(ciphertext);
   }
 
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "createEnv" })])
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "createEnv", reset: true }),
+    ErrorHandlerMW,
+    QuestionMW("createNewEnv"),
+    ConcurrentLockerMW,
+  ])
   async createEnv(inputs: Inputs): Promise<Result<undefined, FxError>> {
-    return this.v3Implement.dispatch(this.createEnv, inputs);
+    return this.createEnvCopyV3(
+      inputs[QuestionNames.NewTargetEnvName]!,
+      inputs[QuestionNames.SourceEnvName]!,
+      inputs.projectPath!
+    );
+  }
+  async createEnvCopyV3(
+    targetEnvName: string,
+    sourceEnvName: string,
+    projectPath: string
+  ): Promise<Result<undefined, FxError>> {
+    let res = await pathUtils.getEnvFilePath(projectPath, sourceEnvName);
+    if (res.isErr()) return err(res.error);
+    const sourceDotEnvFile = res.value;
+
+    res = await pathUtils.getEnvFilePath(projectPath, targetEnvName);
+    if (res.isErr()) return err(res.error);
+    const targetDotEnvFile = res.value;
+    if (!sourceDotEnvFile || !targetDotEnvFile)
+      return err(new YamlFieldMissingError("environmentFolderPath"));
+    if (!(await fs.pathExists(sourceDotEnvFile)))
+      return err(new FileNotFoundError("createEnvCopyV3", sourceDotEnvFile));
+    const source = await fs.readFile(sourceDotEnvFile);
+    const writeStream = fs.createWriteStream(targetDotEnvFile);
+    source
+      .toString()
+      .split(/\r?\n/)
+      .forEach((line) => {
+        const reg = /^([a-zA-Z_][a-zA-Z0-9_]*=)/g;
+        const match = reg.exec(line);
+        if (match) {
+          if (match[1].startsWith("TEAMSFX_ENV=")) {
+            writeStream.write(`TEAMSFX_ENV=${targetEnvName}${os.EOL}`);
+          } else {
+            writeStream.write(`${match[1]}${os.EOL}`);
+          }
+        } else {
+          writeStream.write(`${line.trim()}${os.EOL}`);
+        }
+      });
+
+    writeStream.end();
+    return ok(undefined);
   }
 
   // a phantom migration method for V3
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "phantomMigrationV3" })])
+  @hooks([ErrorContextMW({ component: "FxCore", stage: "phantomMigrationV3", reset: true })])
   async phantomMigrationV3(inputs: Inputs): Promise<Result<undefined, FxError>> {
-    return this.v3Implement.dispatch(this.phantomMigrationV3, inputs);
+    // If the project is invalid or upgraded, the ProjectMigratorMWV3 will not take action.
+    // Check invaliad/upgraded project here before call ProjectMigratorMWV3
+    const projectPath = (inputs.projectPath as string) || "";
+    const version = await getProjectVersionFromPath(projectPath);
+
+    if (version.source === VersionSource.teamsapp) {
+      return err(new NoNeedUpgradeError());
+    } else if (version.source === VersionSource.projectSettings) {
+      const isValid = await checkActiveResourcePlugins(projectPath);
+      if (!isValid) {
+        return err(new InvalidProjectError());
+      }
+    }
+    if (version.source === VersionSource.unknown) {
+      return err(new InvalidProjectError());
+    }
+    return await this.innerMigrationV3(inputs);
+  }
+
+  @hooks([ErrorHandlerMW, ProjectMigratorMWV3])
+  async innerMigrationV3(inputs: Inputs): Promise<Result<undefined, FxError>> {
+    return ok(undefined);
   }
 
   // a project version check
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "projectVersionCheck" })])
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "projectVersionCheck", reset: true }),
+    ErrorHandlerMW,
+  ])
   async projectVersionCheck(inputs: Inputs): Promise<Result<VersionCheckRes, FxError>> {
-    return this.v3Implement.dispatch(this.projectVersionCheck, inputs);
+    const projectPath = (inputs.projectPath as string) || "";
+    if (isValidProjectV3(projectPath) || isValidProjectV2(projectPath)) {
+      const versionInfo = await getProjectVersionFromPath(projectPath);
+      if (!versionInfo.version) {
+        return err(new InvalidProjectError());
+      }
+      const trackingId = await getTrackingIdFromPath(projectPath);
+      const isSupport = getVersionState(versionInfo);
+      // if the project is upgradeable, check whether the project is valid and invalid project should not show upgrade option.
+      if (isSupport === VersionState.upgradeable) {
+        if (!(await checkActiveResourcePlugins(projectPath))) {
+          return err(new InvalidProjectError());
+        }
+      }
+      return ok({
+        currentVersion: versionInfo.version,
+        trackingId,
+        isSupport,
+        versionSource: VersionSource[versionInfo.source],
+      });
+    } else {
+      return err(new InvalidProjectError());
+    }
   }
+
   // apply the given yaml template to current project.
   async apply(
     inputs: Inputs,
@@ -520,37 +1036,119 @@ export class FxCore {
       }
     }
   }
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "preProvisionForVS" })])
+
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "preProvisionForVS" }),
+    ErrorHandlerMW,
+    ProjectMigratorMWV3,
+    EnvLoaderMW(false),
+    ConcurrentLockerMW,
+    ContextInjectorMW,
+  ])
   async preProvisionForVS(inputs: Inputs): Promise<Result<PreProvisionResForVS, FxError>> {
-    return this.v3Implement.dispatch(this.preProvisionForVS, inputs);
+    const context = createDriverContext(inputs);
+    return coordinator.preProvisionForVS(context, inputs as InputsWithProjectPath);
   }
-
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "preCheckYmlAndEnvForVS" })])
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "preCheckYmlAndEnvForVS" }),
+    ErrorHandlerMW,
+    ProjectMigratorMWV3,
+    EnvLoaderMW(false),
+    ConcurrentLockerMW,
+    ContextInjectorMW,
+  ])
   async preCheckYmlAndEnvForVS(inputs: Inputs): Promise<Result<undefined, FxError>> {
-    return this.v3Implement.dispatch(this.preCheckYmlAndEnvForVS, inputs);
+    const context = createDriverContext(inputs);
+    const result = await coordinator.preCheckYmlAndEnvForVS(
+      context,
+      inputs as InputsWithProjectPath
+    );
+    return result;
   }
-
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "publishInDeveloperPortal" })])
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "publishInDeveloperPortal" }),
+    ErrorHandlerMW,
+    ConcurrentLockerMW,
+    ContextInjectorMW,
+  ])
   async publishInDeveloperPortal(inputs: Inputs): Promise<Result<undefined, FxError>> {
-    return this.v3Implement.dispatch(this.publishInDeveloperPortal, inputs);
+    inputs.stage = Stage.publishInDeveloperPortal;
+    const context = createContextV3();
+    return await coordinator.publishInDeveloperPortal(context, inputs as InputsWithProjectPath);
   }
-
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "copilotPluginAddAPI" })])
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "copilotPluginAddAPI" }),
+    ErrorHandlerMW,
+    QuestionMW("copilotPluginAddAPI"),
+    ConcurrentLockerMW,
+  ])
   async copilotPluginAddAPI(inputs: Inputs): Promise<Result<undefined, FxError>> {
-    return this.v3Implement.dispatch(this.copilotPluginAddAPI, inputs);
-  }
+    const operations = inputs[QuestionNames.ApiOperation] as string[];
+    const url = inputs[QuestionNames.ApiSpecLocation] ?? inputs.openAIPluginManifest?.api.url;
+    const manifestPath = inputs[QuestionNames.ManifestPath];
 
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "copilotPluginLoadOpenAIManifest" })])
+    // Get API spec file path from manifest
+    const manifestRes = await manifestUtils._readAppManifest(manifestPath);
+    if (manifestRes.isErr()) {
+      return err(manifestRes.error);
+    }
+    const apiSpecFile = manifestRes.value.composeExtensions![0].apiSpecFile;
+    const outputAPISpecPath = path.join(path.dirname(manifestPath), apiSpecFile!);
+
+    const adaptiveCardFolder = path.join(
+      inputs.projectPath!,
+      AppPackageFolderName,
+      AdaptiveFolderName
+    );
+
+    try {
+      const specParser = new SpecParser(url);
+      await specParser.generate(manifestPath, operations, outputAPISpecPath, adaptiveCardFolder);
+    } catch (e) {
+      let error: FxError;
+      if (e instanceof SpecParserError) {
+        error = convertSpecParserErrorToFxError(e);
+      } else {
+        error = assembleError(e);
+      }
+      return err(error);
+    }
+
+    const message = getLocalizedString(
+      "core.copilot.addAPI.success",
+      operations,
+      inputs.projectPath
+    );
+    await TOOLS.ui.showMessage("info", message, false);
+    return ok(undefined);
+  }
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "copilotPluginLoadOpenAIManifest" }),
+    ErrorHandlerMW,
+  ])
   async copilotPluginLoadOpenAIManifest(
     inputs: Inputs
   ): Promise<Result<OpenAIPluginManifest, FxError>> {
-    return this.v3Implement.dispatch(this.copilotPluginLoadOpenAIManifest, inputs);
+    try {
+      return ok(await OpenAIPluginManifestHelper.loadOpenAIPluginManifest(inputs.domain));
+    } catch (error) {
+      return err(error as FxError);
+    }
   }
-
-  @hooks([ErrorContextMW({ component: "FxCore", stage: "copilotPluginListOperations" })])
+  @hooks([
+    ErrorContextMW({ component: "FxCore", stage: "copilotPluginListOperations" }),
+    ErrorHandlerMW,
+  ])
   async copilotPluginListOperations(
     inputs: Inputs
   ): Promise<Result<ApiOperation[], ErrorResult[]>> {
-    return this.v3Implement.dispatch(this.copilotPluginListOperations, inputs);
+    return await listOperations(
+      createContextV3(),
+      inputs.manifest,
+      inputs.apiSpecUrl,
+      inputs[QuestionNames.ManifestPath],
+      inputs.includeExistingAPIs,
+      inputs.shouldLogWarning
+    );
   }
 }
