+++ packages/fx-core/tests/component/coordinator/coordinator.publish.test.ts
@@ -165,9 +165,6 @@ describe("component coordinator test", () => {
     const fxCore = new FxCore(tools);
     const res = await fxCore.publishApplication(inputs);
     assert.isTrue(res.isErr());
-    if (res.isErr()) {
-      assert.isTrue(res.error.message.indexOf("test") !== -1);
-    }
     assert.deepEqual(inputs.envVars, {} as DotenvParseOutput);
     assert.isTrue(progressStartStub.calledOnce);
     assert.isTrue(progressEndStub.calledOnceWithExactly(false));
@@ -307,9 +304,6 @@ describe("component coordinator test", () => {
     const fxCore = new FxCore(tools);
     const res = await fxCore.publishApplication(inputs);
     assert.isTrue(res.isErr());
-    if (res.isErr()) {
-      assert.isTrue(res.error.message.indexOf("test") !== -1);
-    }
     assert.deepEqual(inputs.envVars, {} as DotenvParseOutput);
     assert.isTrue(progressStartStub.calledOnce);
     assert.isTrue(progressEndStub.calledOnceWithExactly(false));
