+++ packages/fx-core/tests/component/coordinator/coordinator.test.ts
@@ -39,7 +39,6 @@ import { metadataUtil } from "../../../src/component/utils/metadataUtil";
 import { pathUtils } from "../../../src/component/utils/pathUtils";
 import { settingsUtil } from "../../../src/component/utils/settingsUtil";
 import { FxCore } from "../../../src/core/FxCore";
-import { FxCoreV3Implement } from "../../../src/core/FxCoreImplementV3";
 import { setTools } from "../../../src/core/globalVars";
 import * as v3MigrationUtils from "../../../src/core/middleware/utils/v3MigrationUtils";
 import { MissingEnvironmentVariablesError } from "../../../src/error/common";
@@ -290,15 +289,6 @@ describe("component coordinator test", () => {
     assert.isTrue(res.isErr());
   });
 
-  it("buildAadManifest", async () => {
-    sandbox.stub(FxCoreV3Implement.prototype, "buildAadManifest").resolves(ok(undefined));
-    const inputs: Inputs = {
-      platform: Platform.VSCode,
-    };
-    const fxCore = new FxCore(tools);
-    const res1 = await fxCore.buildAadManifest(inputs);
-    assert.isTrue(res1.isOk());
-  });
   it("executeUserTaskNew", async () => {
     sandbox.stub(envUtil, "listEnv").resolves(ok(["dev"]));
     sandbox.stub(envUtil, "readEnv").resolves(ok({}));
